Extended inc.graph.php
Added the ability to specify HTML IDs for created tables (by passing "table.id" in the input array).
For styling purposes, the bars of bar graphs have the class "svgBarGraphBar", lines of line graphs got "svgLineGraphCircle" and "<name of line>svgLineGraphCircle".
It's also possible to output data only in the legend in line graphs. For this the parameter "graph.onlyLegend" was defined, which contains an array of all the lines that shall not be outputted graphically.